requirements.txt
astroid==2.1.0
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.8
django-bootstrap3==11.0.0
django-extensions==2.1.4
django-tastypie==0.14.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.7.5
pydotplus==2.0.2
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.5
requests==2.19.1
six==1.12.0
tabulate==0.8.2
urllib3==1.23
Werkzeug==0.14.1
whitenoise==4.1
wrapt==1.10.11
